[lexical-playground] Refactor: Migrate YouTubePlugin to YouTubeExtension
Converts the legacy React YouTubePlugin into a declarative
YouTubeExtension using defineExtension, registering the YouTubeNode and
INSERT_YOUTUBE_COMMAND through the extension system.
