+++ packages/lexical-playground/src/plugins/AutoEmbedPlugin/index.tsx
@@ -24,7 +24,7 @@ import Button from '../../ui/Button';
 import {DialogActions} from '../../ui/Dialog';
 import {INSERT_FIGMA_COMMAND} from '../FigmaPlugin';
 import {INSERT_TWEET_COMMAND} from '../TwitterExtension';
-import {INSERT_YOUTUBE_COMMAND} from '../YouTubePlugin';
+import {INSERT_YOUTUBE_COMMAND} from '../YouTubeExtension';
 
 interface PlaygroundEmbedConfig extends EmbedConfig {
   // Human readable name of the embedded content e.g. Tweet or Google Map.
