+++ packages/lexical-playground/src/plugins/YouTubeExtension/index.ts
@@ -6,12 +6,13 @@
  *
  */
 
-import type {JSX} from 'react';
-
-import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
 import {$insertNodeToNearestRoot} from '@lexical/utils';
-import {COMMAND_PRIORITY_EDITOR, createCommand, LexicalCommand} from 'lexical';
-import {useEffect} from 'react';
+import {
+  COMMAND_PRIORITY_EDITOR,
+  createCommand,
+  defineExtension,
+  LexicalCommand,
+} from 'lexical';
 
 import {$createYouTubeNode, YouTubeNode} from '../../nodes/YouTubeNode';
 
@@ -19,15 +20,11 @@ export const INSERT_YOUTUBE_COMMAND: LexicalCommand<string> = createCommand(
   'INSERT_YOUTUBE_COMMAND',
 );
 
-export default function YouTubePlugin(): JSX.Element | null {
-  const [editor] = useLexicalComposerContext();
-
-  useEffect(() => {
-    if (!editor.hasNodes([YouTubeNode])) {
-      throw new Error('YouTubePlugin: YouTubeNode not registered on editor');
-    }
-
-    return editor.registerCommand<string>(
+export const YouTubeExtension = defineExtension({
+  name: '@lexical/playground/YouTube',
+  nodes: [YouTubeNode],
+  register: (editor) =>
+    editor.registerCommand<string>(
       INSERT_YOUTUBE_COMMAND,
       (payload) => {
         const youTubeNode = $createYouTubeNode(payload);
@@ -36,8 +33,5 @@ export default function YouTubePlugin(): JSX.Element | null {
         return true;
       },
       COMMAND_PRIORITY_EDITOR,
-    );
-  }, [editor]);
-
-  return null;
-}
+    ),
+});
